Move sqlite-specific state to own module
This change moves the various SqliteBackend-specific items to a sqlite
module, `transact::state::merkle::sql::sqlite`.
